1. 所有接口的分页查询增加返回关联Id文本描述的字段，避免前端手动转换 2. 角色信息修改为支持角色分类管理

+++ src/apis/account.rs
@@ -261,7 +261,7 @@ async fn check_auth_code(key_type: &str, key: &str, code: &str) -> Result<()> {
 
 fn get_user_menus<'a>(menus: &'a [SysMenu], permits: &str) -> Vec<&'a SysMenu> {
     let pbs = bits::string_to_bools(permits);
-    let pbs_len = pbs.len() as i32;
+    let pbs_len = pbs.len() as i16;
     let mut user_menus = Vec::new();
 
     for menu in menus {
